+++ src/index.tsx
@@ -11,9 +11,9 @@ import {
 
 import './styles/style.scss';
 
-import App from './App.jsx';
+import App from './App.tsx';
+import ErrorPage from './containers/ErrorPage.tsx';
 import Main from './containers/Main.jsx';
-import NavBar from './containers/NavBar.jsx';
 import Settings from './containers/Settings.jsx';
 import History from './containers/History.jsx';
 import SocketTest from './components/SocketTest.jsx';
@@ -22,6 +22,7 @@ const router = createBrowserRouter([
   {
     path: '/',
     element: <App />,
+    errorElement: <ErrorPage />,
     children: [
       {
         path: '/',
